Improve json parsing of td object; Fix cast int to double;

+++ lib/src/api/objects/account_ttl.dart
@@ -22,7 +22,7 @@ class AccountTtl extends TdObject {
     }
 
     return AccountTtl(
-      days: json['days'],
+      days: json['days'] as int,
     );
   }
