+++ lib/src/api/objects/authentication_code_type_call.dart
@@ -21,7 +21,7 @@ class AuthenticationCodeTypeCall extends AuthenticationCodeType {
     }
 
     return AuthenticationCodeTypeCall(
-      length: json['length'],
+      length: json['length'] as int,
     );
   }
