+++ lib/src/api/objects/chat_action_uploading_video_note.dart
@@ -20,7 +20,7 @@ class ChatActionUploadingVideoNote extends ChatAction {
     }
 
     return ChatActionUploadingVideoNote(
-      progress: json['progress'],
+      progress: json['progress'] as int,
     );
   }
