+++ lib/src/api/objects/chat_member_status_banned.dart
@@ -25,7 +25,7 @@ class ChatMemberStatusBanned extends ChatMemberStatus {
     }
 
     return ChatMemberStatusBanned(
-      bannedUntilDate: json['banned_until_date'],
+      bannedUntilDate: json['banned_until_date'] as int,
     );
   }
