+++ lib/src/api/objects/inline_keyboard_button_type_user.dart
@@ -21,7 +21,7 @@ class InlineKeyboardButtonTypeUser extends InlineKeyboardButtonType {
     }
 
     return InlineKeyboardButtonTypeUser(
-      userId: json['user_id'],
+      userId: json['user_id'] as int,
     );
   }
