+++ lib/src/api/objects/input_invoice_message.dart
@@ -24,8 +24,8 @@ class InputInvoiceMessage extends InputInvoice {
     }
 
     return InputInvoiceMessage(
-      chatId: json['chat_id'],
-      messageId: json['message_id'],
+      chatId: json['chat_id'] as int,
+      messageId: json['message_id'] as int,
     );
   }
