+++ lib/src/api/objects/message_sender_user.dart
@@ -20,7 +20,7 @@ class MessageSenderUser extends MessageSender {
     }
 
     return MessageSenderUser(
-      userId: json['user_id'],
+      userId: json['user_id'] as int,
     );
   }
