+++ lib/src/api/objects/message_video_chat_ended.dart
@@ -20,7 +20,7 @@ class MessageVideoChatEnded extends MessageContent {
     }
 
     return MessageVideoChatEnded(
-      duration: json['duration'],
+      duration: json['duration'] as int,
     );
   }
