+++ lib/src/api/objects/reset_password_result_pending.dart
@@ -21,7 +21,7 @@ class ResetPasswordResultPending extends ResetPasswordResult {
     }
 
     return ResetPasswordResultPending(
-      pendingResetDate: json['pending_reset_date'],
+      pendingResetDate: json['pending_reset_date'] as int,
     );
   }
