+++ lib/src/api/objects/suggested_action_set_password.dart
@@ -23,7 +23,7 @@ class SuggestedActionSetPassword extends SuggestedAction {
     }
 
     return SuggestedActionSetPassword(
-      authorizationDelay: json['authorization_delay'],
+      authorizationDelay: json['authorization_delay'] as int,
     );
   }
